# openapi.yaml
openapi: 3.0.3
info:
  title: FastAPI React Demo API
  description: 这是一个示例API文档
  version: 1.0.0
  contact:
    name: 技术支持
    email: support@example.com

servers:
  - url: http://localhost:8000
    description: 开发服务器

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        full_name:
          type: string
          example: "张三"
        email:
          type: string
          format: email
          example: "zhangsan@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
    
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "password123"
    
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
    
    SystemStats:
      type: object
      properties:
        cpu_usage:
          type: number
          format: float
          example: 45.2
        memory_usage:
          type: number
          format: float
          example: 62.8
        disk_usage:
          type: number
          format: float
          example: 75.5
        network_io:
          type: object
          properties:
            bytes_sent:
              type: integer
              example: 1024576
            bytes_recv:
              type: integer
              example: 2048576

    ChatMessage:
      type: object
      properties:
        text:
          type: string
          example: "你好，我需要帮助。"
        response:
          type: string
          example: "你好！我很乐意帮助你。请告诉我你的问题。"

paths:
  /api/v1/auth/login:
    post:
      summary: 用户登录
      tags:
        - 认证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 登录失败
  
  /api/v1/users/profile:
    get:
      summary: 获取当前用户信息
      tags:
        - 用户
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功获取用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: 更新用户信息
      tags:
        - 用户
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  example: "李四"
                email:
                  type: string
                  format: email
                  example: "lisi@example.com"
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  
  /api/v1/users/change-password:
    post:
      summary: 修改密码
      tags:
        - 用户
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - old_password
                - new_password
              properties:
                old_password:
                  type: string
                  format: password
                new_password:
                  type: string
                  format: password
      responses:
        '200':
          description: 密码修改成功
        '400':
          description: 旧密码错误
  
  /api/v1/system/stats:
    get:
      summary: 获取系统状态
      tags:
        - 系统
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功获取系统状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStats'
  
  /api/v1/chat/message:
    post:
      summary: 发送聊天消息
      tags:
        - 聊天
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  example: "你好，我需要帮助。"
      responses:
        '200':
          description: 成功获取回复
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage' 